add logic to ensure semaphore is decremented when executions fail to start
